Provide option to select language variants
On Windows, when clicking the "Check" button, provide option to change
language variants -- for instance English (Canada) vs. English (U.S.).
Also, open the modern Editor tool if available, defaulting back to the
old grammar and spell checker if not. Finally, check if a modal dialog
box is open in Word before trying to open the text file.